Helemt Violation Detection/requirements.txt
ultralytics            
opencv-python             
requests                
pytesseract               
Pillow                     
easyocr                    
python-dotenv            
